WFLY-5489 Fix and and add missing transformations for EAP in modcluster subsystem
