+++ mod_cluster/extension/src/main/java/org/wildfly/extension/mod_cluster/ModClusterModel.java
@@ -32,6 +32,7 @@ import org.jboss.as.controller.ModelVersion;
 public enum ModClusterModel {
 
     VERSION_1_4_0(1, 4, 0),
+    VERSION_1_5_0(1, 5, 0),
     VERSION_2_0_0(2, 0, 0),
     VERSION_3_0_0(3, 0, 0),
     ;
