+++ mod_cluster/extension/src/test/java/org/wildfly/extension/mod_cluster/package-info.java
@@ -1,0 +1,39 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2015, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+
+/**
+ * Package containing tests for mod_cluster subsystem.
+ *
+ * Versions overview:
+ *
+ * Server version / model version / schema version
+ * EAP 6.2.0  / 1.4.0 / 1_2
+ * EAP 6.3.0  / 1.5.0 / 1_2
+ * EAP 6.4.0  / 1.5.0 / 1_2
+ * AS  8.0.0  / 2.0.0 / 1_2 (this should have been 2_0!)
+ * AS  9.0.0  / 3.0.0 / 2_0
+ * AS  10.0.0 / 3.0.0 / 2_0
+ *
+ * @author Radoslav Husar
+ * @version Apr 2015
+ */
+package org.wildfly.extension.mod_cluster;
