+++ mod_cluster/extension/src/test/resources/org/wildfly/extension/mod_cluster/subsystem-reject.xml
@@ -21,6 +21,7 @@
   -->
 
 <subsystem xmlns="urn:jboss:domain:modcluster:2.0">
+    <!-- Note: attributes below the blank line are subjected to rejection -->
     <mod-cluster-config advertise="${modcluster.advertise:true}"
                         advertise-security-key="${modcluster.advertise-security-key:mysecurekey!}"
                         advertise-socket="modcluster"
@@ -43,11 +44,11 @@
                         stop-context-timeout="${modcluster.stop-context-timeout:10}"
                         ttl="${modcluster.ttl:1}"
                         worker-timeout="${modcluster.worker-timeout:2}"
-                        session-draining-strategy="${modcluster.session-draining-strategy:ALWAYS}"
-                        status-interval="${modcluster.status-interval:20}"
+
                         proxies="proxy1 proxy2 proxy3"
-        >
-        <simple-load-provider factor="${modcluster.simple-load-provider.factor:15}"/>
+                        session-draining-strategy="${modcluster.session-draining-strategy:ALWAYS}"
+                        status-interval="15"
+            >
         <dynamic-load-provider decay="${modcluster.dynamic-load-provider.decay:2}" history="${modcluster.dynamic-load-provider.history:10}">
             <load-metric capacity="${modcluster.load-metric.capacity:1000.1}" type="sessions" weight="${modcluster.load-metric.weight:2}"/>
             <load-metric capacity="512.2" type="send-traffic" weight="3"/>
