+++ mod_cluster/extension/src/test/resources/org/wildfly/extension/mod_cluster/subsystem-transform.xml
@@ -1,6 +1,6 @@
 <!--
   ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2013, Red Hat, Inc., and individual contributors
+  ~ Copyright 2014, Red Hat, Inc., and individual contributors
   ~ as indicated by the @author tags. See the copyright.txt file in the
   ~ distribution for a full listing of individual contributors.
   ~
@@ -20,7 +20,8 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<subsystem xmlns="urn:jboss:domain:modcluster:1.2">
+<subsystem xmlns="urn:jboss:domain:modcluster:2.0">
+    <!-- Note: attributes below the blank line have values compatible with legacy versions -->
     <mod-cluster-config advertise="${modcluster.advertise:true}"
                         advertise-security-key="${modcluster.advertise-security-key:mysecurekey!}"
                         advertise-socket="modcluster"
@@ -34,7 +35,6 @@
                         max-attempts="${modcluster.max-attempts:10}"
                         node-timeout="${modcluster.node-timeout:123}"
                         ping="${modcluster.ping:10}"
-                        proxy-list="${modcluster.proxy-list:example.com:6666,example.com:6667}"
                         proxy-url="${modcluster.proxy-url:/}"
                         smax="${modcluster.smax:2}"
                         socket-timeout="${modcluster.socket-timeout:20}"
@@ -44,7 +44,9 @@
                         stop-context-timeout="${modcluster.stop-context-timeout:10}"
                         ttl="${modcluster.ttl:1}"
                         worker-timeout="${modcluster.worker-timeout:2}"
-                        session-draining-strategy="${modcluster.session-draining-strategy:ALWAYS}"
+
+                        session-draining-strategy="DEFAULT"
+                        status-interval="10"
             >
         <dynamic-load-provider decay="${modcluster.dynamic-load-provider.decay:2}" history="${modcluster.dynamic-load-provider.history:10}">
             <load-metric capacity="${modcluster.load-metric.capacity:1000.1}" type="sessions" weight="${modcluster.load-metric.weight:2}"/>
