+++ mod_cluster/extension/src/test/resources/org/wildfly/extension/mod_cluster/subsystem_1_0.xml
@@ -21,9 +21,15 @@
   -->
 
 <subsystem xmlns="urn:jboss:domain:modcluster:1.0">
-    <mod-cluster-config proxy-list="example.com:6666" domain="mydomain" advertise="true" advertise-security-key="mysecuritykey"
-                        excluded-contexts="a,b,c" auto-enable-contexts="true" stop-context-timeout="10" socket-timeout="20">
-        <simple-load-provider factor="15"/>
+    <mod-cluster-config advertise="true"
+                        advertise-security-key="mysecuritykey"
+                        auto-enable-contexts="true"
+                        domain="mydomain"
+                        excluded-contexts="a,b,c"
+                        proxy-list="example.com:6666"
+                        stop-context-timeout="10"
+                        socket-timeout="20"
+            >
         <dynamic-load-provider decay="2" history="10">
             <load-metric capacity="1000" type="sessions" weight="2"/>
             <load-metric capacity="512" type="send-traffic" weight="3"/>
@@ -38,11 +44,11 @@
             <custom-load-metric class="org.jboss.modcluster.SomeFakeLoadMetricClass" weight="5"/>
         </dynamic-load-provider>
         <ssl ca-certificate-file="/home/rhusar/client-keystore.jks"
-               ca-revocation-url="/home/rhusar/revocations"
-               certificate-key-file="/home/rhusar/client-keystore.jks"
-               cipher-suite="SSL_DHE_DSS_WITH_3DES_EDE_CBC_SHA,SSL_RSA_WITH_RC4_128_MD5,SSL_RSA_WITH_RC4_128_SHA,SSL_RSA_WITH_3DES_EDE_CBC_SHA"
-               key-alias="mykeyalias"
-               password="mypassword"
-               protocol="TLS"/>
+             ca-revocation-url="/home/rhusar/revocations"
+             certificate-key-file="/home/rhusar/client-keystore.jks"
+             cipher-suite="SSL_DHE_DSS_WITH_3DES_EDE_CBC_SHA,SSL_RSA_WITH_RC4_128_MD5,SSL_RSA_WITH_RC4_128_SHA,SSL_RSA_WITH_3DES_EDE_CBC_SHA"
+             key-alias="mykeyalias"
+             password="mypassword"
+             protocol="TLS"/>
     </mod-cluster-config>
 </subsystem>
