+++ mod_cluster/extension/src/test/resources/org/wildfly/extension/mod_cluster/subsystem_1_1.xml
@@ -21,12 +21,16 @@
   -->
 
 <subsystem xmlns="urn:jboss:domain:modcluster:1.1">
-    <mod-cluster-config proxy-list="example.com:6666" load-balancing-group="mydomain" advertise="true"
+    <mod-cluster-config advertise="true"
                         advertise-security-key="mysecuritykey"
-                        excluded-contexts="a,b,c" auto-enable-contexts="true" stop-context-timeout="10"
+                        auto-enable-contexts="true"
+                        connector="http"
+                        excluded-contexts="a,b,c"
+                        load-balancing-group="mydomain"
+                        proxy-list="example.com:6666"
                         socket-timeout="20"
-                        connector="http">
-        <simple-load-provider factor="15"/>
+                        stop-context-timeout="10"
+            >
         <dynamic-load-provider decay="2" history="10">
             <load-metric capacity="1000" type="sessions" weight="2"/>
             <load-metric capacity="512" type="send-traffic" weight="3"/>
