+++ mod_cluster/extension/src/test/resources/org/wildfly/extension/mod_cluster/subsystem_2_0_simple-load-provider.xml
@@ -21,31 +21,7 @@
   -->
 
 <subsystem xmlns="urn:jboss:domain:modcluster:2.0">
-    <mod-cluster-config advertise="${modcluster.advertise:true}"
-                        advertise-security-key="${modcluster.advertise-security-key:mysecurekey!}"
-                        advertise-socket="modcluster"
-                        auto-enable-contexts="${modcluster.auto-enable-contexts:true}"
-                        balancer="${modcluster.balancer:mybalancer}"
-                        connector="ajp"
-                        excluded-contexts="${modcluster.excluded-contexts:contextA,contextB,contextC}"
-                        flush-packets="${modcluster.flush-packets:true}"
-                        flush-wait="${modcluster.flush-wait:10}"
-                        load-balancing-group="${modcluster.load-balancing-group:mylbgroup}"
-                        max-attempts="${modcluster.max-attempts:10}"
-                        node-timeout="${modcluster.node-timeout:123}"
-                        ping="${modcluster.ping:10}"
-                        proxy-url="${modcluster.proxy-url:/}"
-                        smax="${modcluster.smax:2}"
-                        socket-timeout="${modcluster.socket-timeout:20}"
-                        sticky-session="${modcluster.sticky-session:true}"
-                        sticky-session-force="${modcluster.sticky-session-force:false}"
-                        sticky-session-remove="${modcluster.sticky-session-remove:false}"
-                        stop-context-timeout="${modcluster.stop-context-timeout:10}"
-                        ttl="${modcluster.ttl:1}"
-                        worker-timeout="${modcluster.worker-timeout:2}"
-                        session-draining-strategy="${modcluster.session-draining-strategy:ALWAYS}"
-                        status-interval="10"
-        >
+    <mod-cluster-config connector="ajp">
         <simple-load-provider factor="${modcluster.simple-load-provider.factor:15}"/>
     </mod-cluster-config>
 </subsystem>
